resolved some problems, desired basic functionality achieved, left to do are front-end and additional functionalities that aren't necessary for the app to work as desired save for users

+++ surveyss/src/main/java/surveys/Entities/Questions.java
@@ -17,7 +17,7 @@ public class Questions implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idquestions;
+    private Long idquestions;
 
     @Column(name = "question")
     private String question;
